pyproject.toml
[tool.poetry]
name = "sccca"
version = "0.1.0"
description = ""
authors = ["Harald Vohringer"]

[tool.poetry.dependencies]
python = "^3.8"
pyromaniac = "^0.3.0"
scanpy = "^1.8.2"
pyro-ppl = "^1.8.1"
torch = "^1.11.0"
numpy = "^1.20.0"
pandas = "^1.3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
